+++ tsconfig.json
@@ -17,7 +17,10 @@
       ],
       "@messages/*": [
         "./src/messages/*"
-      ]
+      ],
+      "@service/*": [
+        "./src/service/*"
+      ],
     },
     "outDir": "./dist",
     "rootDir": "./src",
